requirements.txt
PyQt6==6.8.0
pillow==11.1.0
piexif==1.1.3
Send2Trash==1.8.3
